feat(chain): Add Etherlink Testnet (#53)
:wave: this PR is here to add the support for Etherlink Testnet.
Etherlink is a layer 2 blockchain EVM compatible with a well-established
testnet already.

I’m not totally sure about the contributing guidelines to add new
network, so I hope this PR does not come outside of a due process. If it
is the case, please don’t hesitate to redirect me in the right
direction!